Feat : 가게 상세보기 사진 탭 추가 (#89)
