+++ frontend/src/components/RestaurantDetail/RestaurantDetailHeader.tsx
@@ -1,0 +1,9 @@
+const RestaurantDetailHeader = () => (
+  <div className="h-12 flex items-center pl-6 border-b border-gray-100">
+    <span className="text-2xl font-bold tracking-tight text-primary5 font-pretendard">
+      Playce
+    </span>
+  </div>
+);
+
+export default RestaurantDetailHeader;
